Checklist notes inherit from the Table note
Using the table they store the state of checklist entries. These
checklist entries are generated from a standard text note and kept in
sync with a manual Refresh operation.
The power here comes from being able to create a single text note in
ANOTHER Layout, link to it in the current LAYOUT and then hook that link
up to a Checklist... creating a flexible To Do list for major projects
that will expand over time, as new entries are added to the source.